EVM-369 fix architecture dependent `int` type (#1175)
* remove int bound on state sync relayer

* use 64-bit upper bound on removeLowerSnapshots

* upgrade bound of snapshot params to 64-bits

* make chainId type consistent int64 accross codebase

* flag type fix

* lint fix

* lint fix

* increase upper bound of local tx counter

* convert chainid param to uint64

* Fix stake command

* more fixes

* remove explicit type from bitmapSize

* make return of ConvertUnmarshalledInt uint64

---------

+++ command/genesis/params.go
@@ -320,7 +320,7 @@ func (p *genesisParams) initGenesisConfig() error {
 			GasUsed:    command.DefaultGenesisGasUsed,
 		},
 		Params: &chain.Params{
-			ChainID: int(p.chainID),
+			ChainID: int64(p.chainID),
 			Forks:   chain.AllForksEnabled,
 			Engine:  p.consensusEngineConfig,
 		},
